Use constants to create keys for state specific metrics
