+++ server/models/metrics/index.js
@@ -1,7 +1,12 @@
 const US_MO = require("./US_MO");
 const US_PA = require("./US_PA");
+const { METRICS } = require("./shared");
+
+const US_DEMO = { ...US_MO };
 
 exports.default = {
+  US_ND: { ...METRICS },
   US_MO,
   US_PA,
+  US_DEMO,
 };
